CHANGED FILE NAME and added testing
I had to change '-' to '_'  in my file name, I should have done this right at the beginning...
I also added a test. This tests if the sustainability score is calculated correctly by checking if a sustainability score of 10 equals 0 CO2 emission and if a CO2 emission of the average CO2 emission equals a sustainability score of 5.